+++ engine/src/core/time.h
@@ -15,7 +15,7 @@ class ElapsedTime {
 public:
     ElapsedTime() {}
 	void init();
-    int get();
+    float get();
 	operator bool();
 private:
 	std::chrono::time_point<std::chrono::high_resolution_clock> start;
